# Groupe 1 /DADdy/apps/breezy-api/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Breezy API
  version: 1.0.0
  description: >
    Réseau social léger type Twitter/X. Architecture microservices (Go/Gin)
    derrière une API Gateway, données PostgreSQL, authentification JWT.
    Tous les appels passent par la gateway.
servers:
  - url: http://localhost:8080
    description: API Gateway (développement)
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Créer un compte (Fx1)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Compte créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Se connecter (Fx2)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir une session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Nouvelle paire de tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Révoquer un refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Session révoquée
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/me:
    get:
      tags: [Auth]
      summary: Utilisateur courant
      responses:
        '200':
          description: Profil de l'utilisateur authentifié (avec email et rôle)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/request-password-reset:
    post:
      tags: [Auth]
      summary: Demander la réinitialisation du mot de passe
      description: >
        Envoie un lien de réinitialisation par email. Pour éviter l'énumération
        des comptes, la réponse est toujours 202, que l'email existe ou non.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
      responses:
        '202':
          description: Si un compte existe, un lien de réinitialisation a été envoyé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Réinitialiser le mot de passe via un token
      description: >
        Consomme le token reçu par email (usage unique) et définit le nouveau mot
        de passe. Toutes les sessions de l'utilisateur sont révoquées.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Mot de passe réinitialisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Changer son mot de passe (connecté)
      description: >
        Change le mot de passe de l'utilisateur authentifié après vérification du
        mot de passe actuel. Toutes les sessions sont révoquées : il faut se
        reconnecter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Mot de passe changé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /posts:
    post:
      tags: [Posts]
      summary: Publier un message court (Fx3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: VALIDATION_ERROR
            message: content exceeds 280 characters
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHENTICATED
            message: invalid or missing token
...
    Conflict:
      description: Conflit (ressource déjà existante)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: CONFLICT
            message: username already taken
    TooManyRequests:
      description: Trop de requêtes — limite de débit atteinte
      headers:
        Retry-After:
          description: Secondes à attendre avant de réessayer
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: RATE_LIMITED
            message: rate limit exceeded
  schemas:
    PostId:
      type: string
      description: Identifiant opaque du post, préfixé par "pst_".
      pattern: '^pst_[a-f0-9]{16}$'
      example: pst_ab12cd34ef56ab78
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL_ERROR]
        message:
          type: string
...
    UserSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
          description: Identifiant opaque préfixé (ex. "usr_ab12cd34ef56ab78"), émis par l'auth-service.
          example: usr_ab12cd34ef56ab78
        username:
          type: string
        avatarUrl:
          type: string
          nullable: true
    User:
      type: object
      required: [id, username, role, followersCount, followingCount, createdAt]
      properties:
        id:
          type: string
          description: Identifiant opaque préfixé (ex. "usr_ab12cd34ef56ab78"), émis par l'auth-service.
          example: usr_ab12cd34ef56ab78
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
          nullable: true
          description: Renvoyé uniquement lorsque l'utilisateur consulte son propre profil
        displayName:
          type: string
          maxLength: 50
          nullable: true
          description: Nom d'affichage libre, distinct du username
        pronouns:
          type: string
          maxLength: 50
          nullable: true
          description: Pronoms en texte libre (ex. "il/lui", "elle", "they/them")
        bio:
          type: string
          maxLength: 160
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
          description: Image de bannière (en-tête du profil) servie par MinIO
        role:
          type: string
          enum: [user, moderator, admin]
        isPrivate:
          type: boolean
          description: Compte privé — posts/re-posts réservés aux abonnés
        followersCount:
          type: integer
        followingCount:
          type: integer
        isFollowedByMe:
          type: boolean
          description: Contextuel à l'appelant (présent sur GET /users/{id})
        followRequested:
          type: boolean
          description: >
            Contextuel (GET /users/{id}) — true si l'appelant a une demande
            d'abonnement en attente vers ce compte privé
        canViewPosts:
          type: boolean
          description: >
            Contextuel (GET /users/{id}) — false sur un compte privé dont
            l'appelant n'est pas abonné (posts/re-posts masqués)
        createdAt:
          type: string
          format: date-time
    Post:
      type: object
      required: [id, author, content, likesCount, commentsCount, likedByMe, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/PostId'
        author:
          $ref: '#/components/schemas/UserSummary'
        content:
          type: string
          minLength: 1
          maxLength: 280
        tags:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        likedByMe:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      required: [username, email, password, termsAccepted, termsVersion]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        termsAccepted:
          type: boolean
          enum: [true]
          description: Acceptation obligatoire du CLUF en vigueur
        termsVersion:
          type: string
          enum: ['2026-06-12']
          description: Version du CLUF affichée et acceptée par l'utilisateur
    LoginRequest:
      type: object
      required: [password]
      properties:
        identifier:
          type: string
          description: Email ou username. Prioritaire si plusieurs champs sont fournis.
        email:
          type: string
          format: email
          description: Compatibilité ascendante. Utilisé si `identifier` est absent.
        username:
          type: string
          description: Alternative au champ `email`. Utilisé si `identifier` et `email` sont absents.
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          description: JWT signé
        refreshToken:
          type: string
          description: Token opaque stocké côté serveur sous forme de hash
    RequestPasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          description: Token reçu par email (usage unique)
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: Nouveau mot de passe avec au moins une minuscule, une majuscule, un chiffre et un caractère spécial
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
          maxLength: 72
          description: Nouveau mot de passe avec au moins une minuscule, une majuscule, un chiffre et un caractère spécial
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    CreatePostRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 280
        tags:
          type: array
          items:
            type: string
...